change location of handles positions to subviews
Handles' positions was located in plugin options in model, so it was needed to retrieve those values from model to detect collision w/ other handle while moving to avoid overlay of handles & passing of one handle through another. To achieve that I prevoiusly used event emitting and passing other handle position value by presenter from model to specially prepared variable in SliderHandleView.
Today I figured out that these values are not intended to be in options in model because this isn't a part of business logic but a part of presentation logic, so their appropriate place is in view, so I located them in SliderHandleView's params property (params are one common object shared between two SliderHandleViews and passed to them by initialization in SliderView).
So it entailed other changes in SliderScaleView and changes of values types passed to SliderTipView and change in object passed by handleValueChange emit.