🐞 fix: thumb1 could be set to position of thumb2
sliderPointerDown:
add check if thumb keeps distance from another thumb.
This bug could be raised when cursor is clicked on controlContainer right in center between two thumbs, so thumb1 is chosen as closest but allowedPosition for it is chosen the same as position of thumb2, so thumb1 could be set on thumb2 position.