improve findClosestAllowedPosition(), fix bugs
newPosition in pointermove is now initially defined as closest allowed position instead of prevoiusly working with raw percent position value and finding closest allowed position later, before saving and using it on subViews.

add finding of penultimate & max positions to findClosestAllowedPosition():
check if pointer is beyond penultimate position and find closest allowed position using not regular stepSizeInPercents, but difference between penultimate & max positions.
That leads to rounding position in area between penultimate and max precisely to penultimate position if cursor is in left part or to max position if cursor is in right part of that area.
That allows to do not use 'crutch' functions such as isSetToMaxPositionAllowed, isSetToPenultimatePositionAllowed, isOtherThumbOnPenultimatePosition, and even isPointerLessThanHalfStepAwayFromMax and isPointerLessThanHalfStepAwayFromPenultimate, also halfStepFromPenultimateToMax value is no longer needed too.

Also change fixing of values by fixing them right in functions that return values instead of fixing them in caller functions after receiving of a value.