bug fix: step changes relatively to initial value
In Panel element 'step' that sets and receives stepSize of sliderPlugin add set of min property that is equal to step size precision to make value of that input change independently of initial value that is contained in value attribute of element.
Also that adding of min prop (attr) helps to avoid set of 0 value by ↓↑ buttons.